fix: refactor public exports
