+++ ccip-sdk/src/cct/solana/token/constants.ts
@@ -1,0 +1,5 @@
+/** SPL Token authority roles that can be set. */
+export const TOKEN_AUTHORITY_TYPES = {
+  MINT: 'mint',
+  FREEZE: 'freeze',
+} as const
