+++ ccip-sdk/src/cct/solana/token/operations/index.ts
@@ -1,11 +1,4 @@
 export * from './create-token-account.ts'
 export * from './deploy-token.ts'
 export * from './mint-tokens.ts'
-export { SetTokenAuthority } from './set-token-authority.ts'
-export type {
-  ExecuteSetTokenAuthorityParams,
-  ExecuteSetTokenAuthorityResult,
-  GenerateSetTokenAuthorityParams,
-  GenerateSetTokenAuthorityResult,
-  TokenAuthorityType,
-} from './set-token-authority.ts'
+export * from './set-token-authority.ts'
